requirements.txt
streamlit==1.33.0
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
Pillow==10.0.0
pytz==2024.1
